Separated upload module
